Added the use of router view
